fix: resolve #9775 against the landed require path, split inprocess.rs
